Allow to ask if managed user exists
Many management actions only make sense if a user has been selected beforehand.

We created this method to be able to add callbacks to actions that need to have a
user selected in order to avoid exceptions.

For this control I could have updated the existing :only_verified_users method, but I
preferred to keep the two concepts separate.

We need this control as :only_verified_user is not restrictive enough. The reason is
that the :managed_user method used in the :only_verified_user if it does not find a
user it does an initializce (find_or_initialize_by). This causes that when we have "skip_verification" to true, it returns this non-persisted user as "verified".